+++ backend/src/main/java/com/example/backend/infrastructure/repositories/reports/DailyReportRepository.java
@@ -1,10 +1,10 @@
-package com.example.backend.infrastructure.repositories;
+package com.example.backend.infrastructure.repositories.reports;
 
 import static com.example.backend.jooq.tables.JDailyReports.DAILY_REPORTS;
 
-import com.example.backend.domain.entities.DailyReport;
-import com.example.backend.domain.enums.ReportStatus;
-import com.example.backend.domain.repositories.IDailyReportRepository;
+import com.example.backend.domain.reports.DailyReport;
+import com.example.backend.domain.reports.ReportStatus;
+import com.example.backend.domain.reports.IDailyReportRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
